refactor: game_factory to handle create_game & replay_exit methods - WIP
