+++ lib/run_game.rb
@@ -7,7 +7,7 @@ require_relative 'message'
 require_relative 'computer_player'
 require_relative 'human_player'
 require_relative 'game_mode'
-
+require_relative 'game_factory'
 
 def start_game
   board = Board.new
@@ -16,30 +16,22 @@ def start_game
   display = Display.new(message, board, validate_response)
   game_mode = GameMode.new(display, board)
 
-  display.print_welcome
-  player1 = game_mode.get_player1(game_mode.choose_game_mode)
-  player2 = HumanPlayer.new('O', 'Human', display)
-  game = Game.new(board, display, player1, player2)
+  game_factory = GameFactory.new(display, game_mode)
+  game_factory.create_game
+  # display.print_welcome
+  # player1 = game_mode.get_player1(game_mode.choose_game_mode)
+  # player2 = HumanPlayer.new('O', 'Human', display)
+  # game = Game.new(board, display, player1, player2)
 
-  display.print_game_starting
-  display.print_enter_num
-  game.start_game
-  display.print_play_again_exit
-  display.play_exit_choice
-  replay_or_exit(display.validate_play_again_choice)
-  
-end
+  # display.print_game_starting
+  # display.print_enter_num
 
-def replay_or_exit(play_again_choice)
-  if play_again_choice == 1 
-    start_game 
-  else
-     print "\nThanks for playing Tic Tac Toe!\n\n"
-  end
+  # game.start_game
+
+  # display.print_play_again_exit
+  # display.play_exit_choice
+
+  # replay_or_exit(display.validate_play_again_choice)
 end
 
 start_game
-
-
-
-
