+++ spec/game_factor_spec.rb
@@ -1,0 +1,1 @@
+require 'game_factory'
