+++ spec/game_mode_spec.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 require 'game_mode'
 require 'board'
 require 'display'
